requirements.txt
flask
pillow
anthropic
python-dotenv
